Fix for newer Encode module(>= 2.53)
In newer Encode, we cannot pass decoded string to decode methods
such as `decoded_utf8` and we get error if we pass decoded string.
Some tests are failed with newer Encode because of this.

I suppose that `visual_length` must take string(not octets) if user
uses `bin/md2inao.pl` so we need not to decode argument of `visual_length`.